backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
motor==3.3.1
python-dotenv==1.1.0
slack-sdk==3.27.1
pydantic==2.11.6
qrcode[pil]==8.2
pillow==11.2.1
apscheduler==3.11.0
bcrypt==4.3.0
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
requests==2.32.4
pymongo==4.5.0 